Plan oversized module decomposition
