+++ src/gpt_trader/monitoring/health_checks.py
@@ -13,10 +13,17 @@ from __future__ import annotations
 
 import time
 from collections.abc import Callable
-from dataclasses import dataclass
 from functools import partial
-from typing import TYPE_CHECKING, Any, Literal, cast
+from typing import TYPE_CHECKING, Any, cast
 
+from gpt_trader.monitoring.health_check_planning import (
+    HealthCheckCycleError,
+    HealthCheckDependencyError,
+    HealthCheckDescriptor,
+    HealthCheckOutcome,
+    HealthCheckPlanError,
+    HealthCheckPlanner,
+)
 from gpt_trader.monitoring.interfaces import HealthCheckDependency, HealthCheckResult
 from gpt_trader.monitoring.metrics_collector import record_counter
 from gpt_trader.monitoring.profiling import profile_span
@@ -46,136 +53,7 @@ TICKER_CACHE_UNAVAILABLE_COUNTER = "gpt_trader_ticker_cache_unavailable_total"
 TICKER_STALE_SYMBOLS_COUNTER = "gpt_trader_ticker_stale_symbols_total"
 
 
-HealthCheckOutcome = HealthCheckResult
-HealthCheckMode = Literal["blocking", "fast"]
-
 HealthCheckCallable = Callable[[], HealthCheckResult | tuple[bool, dict[str, Any]]]
-
-
-class HealthCheckPlanError(RuntimeError):
-    """Base error for health check planning failures."""
-
-
-class HealthCheckDependencyError(HealthCheckPlanError):
-    """Raised when required dependencies are missing."""
-
-    def __init__(self, missing_dependencies: dict[str, tuple[str, ...]]) -> None:
-        self.missing_dependencies = missing_dependencies
-        missing_details = "; ".join(
-            f"{check} requires {', '.join(dependencies)}"
-            for check, dependencies in sorted(missing_dependencies.items())
-        )
-        super().__init__(f"Missing required health check dependencies: {missing_details}")
-
-
-class HealthCheckCycleError(HealthCheckPlanError):
-    """Raised when a dependency cycle is detected."""
-
-    def __init__(self, cycle: tuple[str, ...]) -> None:
-        self.cycle = cycle
-        cycle_text = " -> ".join(cycle) if cycle else "unknown cycle"
-        super().__init__(f"Health check dependency cycle detected: {cycle_text}")
-
-
-@dataclass(frozen=True)
-class HealthCheckDescriptor:
-    """Descriptor for a registered health check."""
-
-    name: str
-    mode: HealthCheckMode
-    run: Callable[[], HealthCheckOutcome]
-    dependencies: tuple[HealthCheckDependency, ...] = ()
-
-
-class HealthCheckPlanner:
-    """Build deterministic execution order for health checks."""
-
-    def __init__(self, checks: tuple[HealthCheckDescriptor, ...]) -> None:
-        self._checks = checks
-        self._checks_by_name = {check.name: check for check in checks}
-
-    def build_order(self) -> tuple[HealthCheckDescriptor, ...]:
-        """Resolve dependencies and return ordered health checks."""
-        if not self._checks_by_name:
-            return ()
-
-        dependency_graph, missing_required = self._build_dependency_graph()
-        if missing_required:
-            raise HealthCheckDependencyError(
-                {name: tuple(sorted(missing)) for name, missing in sorted(missing_required.items())}
-            )
-
-        ordered_names = self._topological_sort(dependency_graph)
-        return tuple(self._checks_by_name[name] for name in ordered_names)
-
-    def _build_dependency_graph(
-        self,
-    ) -> tuple[dict[str, set[str]], dict[str, list[str]]]:
-        dependency_graph: dict[str, set[str]] = {name: set() for name in self._checks_by_name}
-        missing_required: dict[str, list[str]] = {}
-
-        for check in self._checks_by_name.values():
-            for dependency in check.dependencies:
-                if dependency.name in self._checks_by_name:
-                    dependency_graph[check.name].add(dependency.name)
-                elif dependency.required:
-                    missing_required.setdefault(check.name, []).append(dependency.name)
-
-        return dependency_graph, missing_required
-
-    def _topological_sort(self, dependency_graph: dict[str, set[str]]) -> tuple[str, ...]:
-        remaining_dependencies = {
-            name: set(dependencies) for name, dependencies in dependency_graph.items()
-        }
-        ready = sorted(name for name, deps in remaining_dependencies.items() if not deps)
-        ordered: list[str] = []
-
-        while ready:
-            name = ready.pop(0)
-            ordered.append(name)
-            for dependent, dependencies in remaining_dependencies.items():
-                if name in dependencies:
-                    dependencies.remove(name)
-                    if not dependencies and dependent not in ordered and dependent not in ready:
-                        ready.append(dependent)
-                        ready.sort()
-
-        if len(ordered) != len(remaining_dependencies):
-            cycle = self._find_cycle(dependency_graph)
-            raise HealthCheckCycleError(cycle)
-
-        return tuple(ordered)
-
-    def _find_cycle(self, dependency_graph: dict[str, set[str]]) -> tuple[str, ...]:
-        visiting: set[str] = set()
-        visited: set[str] = set()
-        stack: list[str] = []
-
-        def visit(node: str) -> tuple[str, ...] | None:
-            visiting.add(node)
-            stack.append(node)
-            for neighbor in dependency_graph.get(node, set()):
-                if neighbor in visiting:
-                    if neighbor in stack:
-                        start_index = stack.index(neighbor)
-                        return tuple(stack[start_index:] + [neighbor])
-                    return (neighbor, node, neighbor)
-                if neighbor not in visited:
-                    cycle_path = visit(neighbor)
-                    if cycle_path:
-                        return cycle_path
-            visiting.remove(node)
-            visited.add(node)
-            stack.pop()
-            return None
-
-        for node in sorted(dependency_graph):
-            if node not in visited:
-                cycle = visit(node)
-                if cycle:
-                    return cycle
-
-        return ()
 
 
 def _coerce_health_check_result(
